Replace typings with npm-managed header files
Also change all imports to not use file extensions (I don't know how
that worked before) and adjust some types in light of new header files.

This also has a bunch of unrelated changes because I ended up in a mess
due to not understanding the intricacies of getting typescript and react
to communicate and thus got myself into a mess :)

Also commented out the test. This is kind of terrible. I'm not sure how
to use TestUtils effectively, next step is to learn that.